requirements.txt
bittensor==6.9.3
cryptography==42.0.0
Pillow==10.1.0
Requests==2.31.0
tqdm==4.66.1
httpx==0.27.0
prometheus_fastapi_instrumentator==6.0.0
pymongo==4.7.3
slowapi==0.1.9
bcrypt==4.1.3
transformers
jinja2==3.1.0
python-dotenv==1.0.1
stripe